main menu merged into main copy, scene loader works properly

+++ Blades_and_Bullets/Assets/Scripts/Scripts-Ethan.H/Enemy.cs
@@ -34,7 +34,7 @@ public class WaveEnemy : MonoBehaviour
 
     [SerializeField]
     private int maxHP;
-    private int currentHP;
+    private float currentHP;
     //[SerializeField]
     //private int pointDropCount;
     //[SerializeField]
